RF: Move the logging to utils.
